Rename existing code from piffle.iiif to piffle.image
